Update prettier-plugin-solidity (#1294)

+++ core/contracts/common/implementation/ExpandedERC20.sol
@@ -25,14 +25,15 @@ contract ExpandedERC20 is ExpandedIERC20, ERC20, MultiRole {
      * @param _tokenSymbol The ticker abbreviation of the name. Ideally < 5 chars.
      * @param _tokenDecimals The number of decimals to define token precision.
      */
-    constructor(string memory _tokenName, string memory _tokenSymbol, uint8 _tokenDecimals)
-        public
-        ERC20(_tokenName, _tokenSymbol)
-    {
+    constructor(
+        string memory _tokenName,
+        string memory _tokenSymbol,
+        uint8 _tokenDecimals
+    ) public ERC20(_tokenName, _tokenSymbol) {
         _setupDecimals(_tokenDecimals);
-        _createExclusiveRole(uint(Roles.Owner), uint(Roles.Owner), msg.sender);
-        _createSharedRole(uint(Roles.Minter), uint(Roles.Owner), new address[](0));
-        _createSharedRole(uint(Roles.Burner), uint(Roles.Owner), new address[](0));
+        _createExclusiveRole(uint256(Roles.Owner), uint256(Roles.Owner), msg.sender);
+        _createSharedRole(uint256(Roles.Minter), uint256(Roles.Owner), new address[](0));
+        _createSharedRole(uint256(Roles.Burner), uint256(Roles.Owner), new address[](0));
     }
 
     /**
@@ -41,10 +42,12 @@ contract ExpandedERC20 is ExpandedIERC20, ERC20, MultiRole {
      * @param value amount of tokens to mint.
      * @return True if the mint succeeded, or False.
      */
-
-    // TODO(#969) Remove once prettier-plugin-solidity can handle the "override" keyword
-    // prettier-ignore
-    function mint(address recipient, uint256 value) external override onlyRoleHolder(uint(Roles.Minter)) returns (bool) {
+    function mint(address recipient, uint256 value)
+        external
+        override
+        onlyRoleHolder(uint256(Roles.Minter))
+        returns (bool)
+    {
         _mint(recipient, value);
         return true;
     }
@@ -53,10 +56,7 @@ contract ExpandedERC20 is ExpandedIERC20, ERC20, MultiRole {
      * @dev Burns `value` tokens owned by `msg.sender`.
      * @param value amount of tokens to burn.
      */
-
-    // TODO(#969) Remove once prettier-plugin-solidity can handle the "override" keyword
-    // prettier-ignore
-    function burn(uint256 value) external override onlyRoleHolder(uint(Roles.Burner)) {
+    function burn(uint256 value) external override onlyRoleHolder(uint256(Roles.Burner)) {
         _burn(msg.sender, value);
     }
 }
