+++ core/contracts/financial-templates/implementation/SyntheticToken.sol
@@ -15,10 +15,11 @@ contract SyntheticToken is ExpandedERC20 {
      * @param tokenSymbol The ticker abbreviation of the name. Ideally < 5 chars.
      * @param tokenDecimals The number of decimals to define token precision.
      */
-    constructor(string memory tokenName, string memory tokenSymbol, uint8 tokenDecimals)
-        public
-        ExpandedERC20(tokenName, tokenSymbol, tokenDecimals)
-    {}
+    constructor(
+        string memory tokenName,
+        string memory tokenSymbol,
+        uint8 tokenDecimals
+    ) public ExpandedERC20(tokenName, tokenSymbol, tokenDecimals) {}
 
     /**
      * @notice Add Minter role to account.
@@ -26,7 +27,7 @@ contract SyntheticToken is ExpandedERC20 {
      * @param account The address to which the Minter role is added.
      */
     function addMinter(address account) external {
-        addMember(uint(Roles.Minter), account);
+        addMember(uint256(Roles.Minter), account);
     }
 
     /**
@@ -35,7 +36,7 @@ contract SyntheticToken is ExpandedERC20 {
      * @param account The address from which the Minter role is removed.
      */
     function removeMinter(address account) external {
-        removeMember(uint(Roles.Minter), account);
+        removeMember(uint256(Roles.Minter), account);
     }
 
     /**
@@ -44,7 +45,7 @@ contract SyntheticToken is ExpandedERC20 {
      * @param account The address to which the Burner role is added.
      */
     function addBurner(address account) external {
-        addMember(uint(Roles.Burner), account);
+        addMember(uint256(Roles.Burner), account);
     }
 
     /**
@@ -53,7 +54,7 @@ contract SyntheticToken is ExpandedERC20 {
      * @param account The address from which the Burner role is removed.
      */
     function removeBurner(address account) external {
-        removeMember(uint(Roles.Burner), account);
+        removeMember(uint256(Roles.Burner), account);
     }
 
     /**
@@ -62,7 +63,7 @@ contract SyntheticToken is ExpandedERC20 {
      * @param account The new holder of the Owner role.
      */
     function resetOwner(address account) external {
-        resetMember(uint(Roles.Owner), account);
+        resetMember(uint256(Roles.Owner), account);
     }
 
     /**
@@ -71,7 +72,7 @@ contract SyntheticToken is ExpandedERC20 {
      * @return bool True if the provided account is a Minter.
      */
     function isMinter(address account) public view returns (bool) {
-        return holdsRole(uint(Roles.Minter), account);
+        return holdsRole(uint256(Roles.Minter), account);
     }
 
     /**
@@ -80,6 +81,6 @@ contract SyntheticToken is ExpandedERC20 {
      * @return bool True if the provided account is a Burner.
      */
     function isBurner(address account) public view returns (bool) {
-        return holdsRole(uint(Roles.Burner), account);
+        return holdsRole(uint256(Roles.Burner), account);
     }
 }
