+++ core/contracts/oracle/implementation/Governor.sol
@@ -17,7 +17,7 @@ import "@openzeppelin/contracts/utils/Address.sol";
  * @title Takes proposals for certain governance actions and allows UMA token holders to vote on them.
  */
 contract Governor is MultiRole, Testable {
-    using SafeMath for uint;
+    using SafeMath for uint256;
     using Address for address;
 
     /****************************************
@@ -60,14 +60,18 @@ contract Governor is MultiRole, Testable {
      * @param _timerAddress Contract that stores the current time in a testing environment.
      * Must be set to 0x0 for production environments that use live time.
      */
-    constructor(address _finderAddress, uint _startingId, address _timerAddress) public Testable(_timerAddress) {
+    constructor(
+        address _finderAddress,
+        uint256 _startingId,
+        address _timerAddress
+    ) public Testable(_timerAddress) {
         finder = FinderInterface(_finderAddress);
-        _createExclusiveRole(uint(Roles.Owner), uint(Roles.Owner), msg.sender);
-        _createExclusiveRole(uint(Roles.Proposer), uint(Roles.Owner), msg.sender);
+        _createExclusiveRole(uint256(Roles.Owner), uint256(Roles.Owner), msg.sender);
+        _createExclusiveRole(uint256(Roles.Proposer), uint256(Roles.Owner), msg.sender);
 
         // Ensure the startingId is not set unreasonably high to avoid it being set such that new proposals overwrite
         // other storage slots in the contract.
-        uint maxStartingId = 10**18;
+        uint256 maxStartingId = 10**18;
         require(_startingId <= maxStartingId, "Cannot set startingId larger than 10^18");
 
         // This just sets the initial length of the array to the startingId since modifying length directly has been
@@ -92,7 +96,7 @@ contract Governor is MultiRole, Testable {
      * Note: this method must be public because of a solidity limitation that
      * disallows structs arrays to be passed to external functions.
      */
-    function propose(Transaction[] memory transactions) public onlyRoleHolder(uint(Roles.Proposer)) {
+    function propose(Transaction[] memory transactions) public onlyRoleHolder(uint256(Roles.Proposer)) {
         uint256 id = proposals.length;
         uint256 time = getCurrentTime();
 
@@ -165,7 +169,7 @@ contract Governor is MultiRole, Testable {
      * @notice Gets the total number of proposals (includes executed and non-executed).
      * @return uint256 representing the current number of proposals.
      */
-    function numProposals() external view returns (uint) {
+    function numProposals() external view returns (uint256) {
         return proposals.length;
     }
 
@@ -183,7 +187,11 @@ contract Governor is MultiRole, Testable {
      *      PRIVATE GETTERS AND FUNCTIONS   *
      ****************************************/
 
-    function _executeCall(address to, uint256 value, bytes memory data) private returns (bool) {
+    function _executeCall(
+        address to,
+        uint256 value,
+        bytes memory data
+    ) private returns (bool) {
         // Mostly copied from:
         // solhint-disable-next-line max-line-length
         // https://github.com/gnosis/safe-contracts/blob/59cfdaebcd8b87a0a32f87b50fead092c10d3a05/contracts/base/Executor.sol#L23-L31
@@ -256,7 +264,11 @@ contract Governor is MultiRole, Testable {
     // 1. If the resulting UTF-8 is larger than 32 characters, then only the first 32 characters will be represented
     //    by the bytes32 output.
     // 2. If `prefix` has more characters than `prefixLength`, the function will produce an invalid result.
-    function _addPrefix(bytes32 input, bytes32 prefix, uint256 prefixLength) internal pure returns (bytes32) {
+    function _addPrefix(
+        bytes32 input,
+        bytes32 prefix,
+        uint256 prefixLength
+    ) internal pure returns (bytes32) {
         // Downshift `input` to open space at the "front" of the bytes32
         bytes32 shiftedInput = input >> (prefixLength * 8);
         return shiftedInput | prefix;
