+++ core/contracts/oracle/implementation/Registry.sol
@@ -13,7 +13,7 @@ import "@openzeppelin/contracts/math/SafeMath.sol";
  * to register new financial contracts and stores party members of a financial contract.
  */
 contract Registry is RegistryInterface, MultiRole {
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
     /****************************************
      *    INTERNAL VARIABLES AND STORAGE    *
@@ -37,7 +37,7 @@ contract Registry is RegistryInterface, MultiRole {
     struct Party {
         address[] contracts; // Each financial contract address is stored in this array.
         // The address of each financial contract is mapped to its index for constant time look up and deletion.
-        mapping(address => uint) contractIndex;
+        mapping(address => uint256) contractIndex;
     }
 
     // Array of all contracts that are approved to use the UMA Oracle.
@@ -61,9 +61,9 @@ contract Registry is RegistryInterface, MultiRole {
      * @notice Construct the Registry contract.
      */
     constructor() public {
-        _createExclusiveRole(uint(Roles.Owner), uint(Roles.Owner), msg.sender);
+        _createExclusiveRole(uint256(Roles.Owner), uint256(Roles.Owner), msg.sender);
         // Start with no contract creators registered.
-        _createSharedRole(uint(Roles.ContractCreator), uint(Roles.Owner), new address[](0));
+        _createSharedRole(uint256(Roles.ContractCreator), uint256(Roles.Owner), new address[](0));
     }
 
     /****************************************
@@ -76,12 +76,10 @@ contract Registry is RegistryInterface, MultiRole {
      * @param parties array of addresses who become parties in the contract.
      * @param contractAddress address of the contract against which the parties are registered.
      */
-    // TODO(#969) Remove once prettier-plugin-solidity can handle the "override" keyword
-    // prettier-ignore
     function registerContract(address[] calldata parties, address contractAddress)
         external
         override
-        onlyRoleHolder(uint(Roles.ContractCreator))
+        onlyRoleHolder(uint256(Roles.ContractCreator))
     {
         FinancialContract storage financialContract = contractMap[contractAddress];
         require(contractMap[contractAddress].valid == Validity.Invalid, "Can only register once");
@@ -106,8 +104,6 @@ contract Registry is RegistryInterface, MultiRole {
      * @dev msg.sender will be used to determine the contract that this party is added to.
      * @param party new party for the calling contract.
      */
-    // TODO(#969) Remove once prettier-plugin-solidity can handle the "override" keyword
-    // prettier-ignore
     function addPartyToContract(address party) external override {
         address contractAddress = msg.sender;
         require(contractMap[contractAddress].valid == Validity.Valid, "Can only add to valid contract");
@@ -120,13 +116,11 @@ contract Registry is RegistryInterface, MultiRole {
      * @dev msg.sender will be used to determine the contract that this party is removed from.
      * @param partyAddress address to be removed from the calling contract.
      */
-    // TODO(#969) Remove once prettier-plugin-solidity can handle the "override" keyword
-    // prettier-ignore
     function removePartyFromContract(address partyAddress) external override {
         address contractAddress = msg.sender;
         Party storage party = partyMap[partyAddress];
         uint256 numberOfContracts = party.contracts.length;
-        
+
         require(numberOfContracts != 0, "Party has no contracts");
         require(contractMap[contractAddress].valid == Validity.Valid, "Remove only from valid contract");
         require(isPartyMemberOfContract(partyAddress, contractAddress), "Can only remove existing party");
@@ -160,8 +154,6 @@ contract Registry is RegistryInterface, MultiRole {
      * @param contractAddress address of the financial contract.
      * @return bool indicates whether the contract is registered.
      */
-    // TODO(#969) Remove once prettier-plugin-solidity can handle the "override" keyword
-    // prettier-ignore
     function isContractRegistered(address contractAddress) external override view returns (bool) {
         return contractMap[contractAddress].valid == Validity.Valid;
     }
@@ -171,8 +163,6 @@ contract Registry is RegistryInterface, MultiRole {
      * @param party address of the party.
      * @return an array of the contracts the party is registered to.
      */
-    // TODO(#969) Remove once prettier-plugin-solidity can handle the "override" keyword
-    // prettier-ignore
     function getRegisteredContracts(address party) external override view returns (address[] memory) {
         return partyMap[party].contracts;
     }
@@ -181,8 +171,6 @@ contract Registry is RegistryInterface, MultiRole {
      * @notice Returns all registered contracts.
      * @return all registered contract addresses within the system.
      */
-    // TODO(#969) Remove once prettier-plugin-solidity can handle the "override" keyword
-    // prettier-ignore
     function getAllRegisteredContracts() external override view returns (address[] memory) {
         return registeredContracts;
     }
@@ -193,8 +181,6 @@ contract Registry is RegistryInterface, MultiRole {
      * @param contractAddress address to check against the party.
      * @return bool indicating if the address is a party of the contract.
      */
-    // TODO(#969) Remove once prettier-plugin-solidity can handle the "override" keyword
-    // prettier-ignore
     function isPartyMemberOfContract(address party, address contractAddress) public override view returns (bool) {
         uint256 index = partyMap[party].contractIndex[contractAddress];
         return partyMap[party].contracts.length > index && partyMap[party].contracts[index] == contractAddress;
