+++ core/contracts/tokenized-derivative/ManualPriceFeed.sol
@@ -11,7 +11,7 @@ import "./PriceFeedInterface.sol";
  * @title Implementation of PriceFeedInterface with the ability to manually push prices.
  */
 contract ManualPriceFeed is PriceFeedInterface, Withdrawable, Testable {
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
     // A single price update.
     struct PriceTick {
@@ -32,19 +32,20 @@ contract ManualPriceFeed is PriceFeedInterface, Withdrawable, Testable {
     enum Roles { Governance, Writer, Withdraw }
 
     constructor(address _timerAddress) public Testable(_timerAddress) {
-        _createExclusiveRole(uint(Roles.Governance), uint(Roles.Governance), msg.sender);
-        _createExclusiveRole(uint(Roles.Writer), uint(Roles.Governance), msg.sender);
-        _createWithdrawRole(uint(Roles.Withdraw), uint(Roles.Governance), msg.sender);
+        _createExclusiveRole(uint256(Roles.Governance), uint256(Roles.Governance), msg.sender);
+        _createExclusiveRole(uint256(Roles.Writer), uint256(Roles.Governance), msg.sender);
+        _createWithdrawRole(uint256(Roles.Withdraw), uint256(Roles.Governance), msg.sender);
     }
 
     /**
      * @notice Adds a new price to the series for a given identifier.
      * @dev The pushed publishTime must be later than the last time pushed so far.
      */
-    function pushLatestPrice(bytes32 identifier, uint256 publishTime, int256 newPrice)
-        external
-        onlyRoleHolder(uint(Roles.Writer))
-    {
+    function pushLatestPrice(
+        bytes32 identifier,
+        uint256 publishTime,
+        int256 newPrice
+    ) external onlyRoleHolder(uint256(Roles.Writer)) {
         require(publishTime <= getCurrentTime().add(BLOCK_TIMESTAMP_TOLERANCE));
         require(publishTime > prices[identifier].timestamp);
         prices[identifier] = PriceTick(publishTime, newPrice);
@@ -54,14 +55,10 @@ contract ManualPriceFeed is PriceFeedInterface, Withdrawable, Testable {
     /**
      * @notice Whether this feed has ever published any prices for this identifier.
      */
-    // TODO(#969) Remove once prettier-plugin-solidity can handle the "override" keyword
-    // prettier-ignore
     function isIdentifierSupported(bytes32 identifier) external override view returns (bool isSupported) {
         isSupported = _isIdentifierSupported(identifier);
     }
 
-    // TODO(#969) Remove once prettier-plugin-solidity can handle the "override" keyword
-    // prettier-ignore
     function latestPrice(bytes32 identifier) external override view returns (uint256 publishTime, int256 price) {
         require(_isIdentifierSupported(identifier));
         publishTime = prices[identifier].timestamp;
